+++ sdk/ai/Azure.AI.Projects/CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Features Added
 
 ### Breaking Changes
+* Method argument name changes:
+  * In Datasets methods `PendingUpload` and `PendingUploadAsync`, argument `body` was replaced with `pendingUploadRequest`
+
 
 * Removing `GetChatCompletionsClient`, `GetEmbeddingsClient`, and `GetImageEmbeddingsClient` methods from `AIProjectClient`. The Inference client should be used directly instead.