+++ sdk/resources/Azure.ResourceManager.Resources/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Release History
 
-## 1.10.0-beta.1 (2025-06-06)
+## 1.11.0 (2025-06-23)
+
+### Features Added
+
+- Bump `DeploymentScript` api-version to `2023-08-01`.
+
+## 1.10.0 (2025-06-06)
 
 ### Features Added